+++ src/main/java/org/example/princess_group/domain/auth/service/AuthService.java
@@ -1,9 +1,9 @@
 package org.example.princess_group.domain.auth.service;
 
 import java.util.List;
-import org.example.princess_group.domain.auth.dto.AuthInfo;
-import org.example.princess_group.domain.auth.entity.DomainType;
-import org.example.princess_group.domain.auth.entity.ServiceAuthority;
+import org.example.princess_group.global.dto.AuthInfo;
+import org.example.princess_group.global.entity.DomainType;
+import org.example.princess_group.global.entity.ServiceAuthority;
 
 public interface AuthService {
