+++ src/main/java/org/example/princess_group/global/entity/DomainType.java
@@ -1,4 +1,4 @@
-package org.example.princess_group.domain.auth.entity;
+package org.example.princess_group.global.entity;
 
 public enum DomainType {
     BOARD, LIST, CARD, COMMENT
